refactor(types): migrate to typescript project references
