+++ src/app/composables/url.ts
@@ -1,3 +1,6 @@
+export const useSiteUrl = () =>
+  getSiteUrl(useRuntimeConfig().public.i18n.baseUrl)
+
 export const useTusdFilesUrl = () => {
   const { siteUrlTyped: siteUrl } = useSiteUrl()
