+++ src/app/utils/authentication.ts
