+++ src/app/utils/lifecycle.ts
@@ -1,0 +1,8 @@
+import { consola } from 'consola'
+
+import type { NuxtError } from '#app'
+
+export const showAppError = (error: string | Partial<NuxtError>) => {
+  showError(error)
+  consola.error(error)
+}
