+++ src/server/utils/jwt.ts
@@ -1,6 +1,38 @@
 import { setCookie } from 'h3'
 import type { H3Event } from 'h3'
 
+export const setJwtCookie = ({
+  event,
+  jwt,
+  runtimeConfig,
+}: {
+  event: H3Event
+  jwt: string
+  runtimeConfig: ReturnType<typeof useRuntimeConfig>
+}) => {
+  const dateEpoch = new Date(0)
+  const dateInAMonth = new Date(Date.now() + 86400 * 1000 * 31)
+
+  if (!runtimeConfig.public.i18n.baseUrl) {
+    return throwError({
+      statusCode: 500,
+      statusMessage: 'Site URL is not defined in the runtime configuration.',
+    })
+  }
+
+  const siteUrl = new URL(runtimeConfig.public.i18n.baseUrl as string) // TODO: remove typecast in @nuxtjs/i18n v11
+  const isHttps = siteUrl.protocol === 'https:'
+  const jwtCookieName = JWT_NAME({ isHttps })
+
+  setCookie(event, jwtCookieName, jwt, {
+    expires: jwt ? dateInAMonth : dateEpoch,
+    httpOnly: true,
+    // path: '/',
+    sameSite: 'lax', // Cannot be 'strict' to allow authentications after clicking on links within webmailers.
+    secure: isHttps,
+  })
+}
+
 export const useJsonWebToken = async () => {
   const event = useEvent()
   const runtimeConfig = useRuntimeConfig()
@@ -56,34 +88,16 @@ export const useJsonWebToken = async () => {
   }
 }
 
-export const setJwtCookie = ({
-  event,
-  jwt,
-  runtimeConfig,
-}: {
-  event: H3Event
-  jwt: string
-  runtimeConfig: ReturnType<typeof useRuntimeConfig>
-}) => {
-  const dateEpoch = new Date(0)
-  const dateInAMonth = new Date(Date.now() + 86400 * 1000 * 31)
+export const useJwtName = () => getJwtName(useSiteUrl().siteUrlTyped)
 
-  if (!runtimeConfig.public.i18n.baseUrl) {
-    return throwError({
-      statusCode: 500,
-      statusMessage: 'Site URL is not defined in the runtime configuration.',
-    })
+export const useJwtPublicKey = async () => {
+  const runtimeConfig = useRuntimeConfig()
+
+  if (runtimeConfig.public.vio.stagingHost) {
+    return await $fetch<string>(
+      `https://${runtimeConfig.public.vio.stagingHost}/api/service/postgraphile/jwt-public-key`,
+    )
+  } else {
+    return runtimeConfig.public.vio.auth.jwt.publicKey
   }
-
-  const siteUrl = new URL(runtimeConfig.public.i18n.baseUrl as string) // TODO: remove typecast in @nuxtjs/i18n v11
-  const isHttps = siteUrl.protocol === 'https:'
-  const jwtCookieName = JWT_NAME({ isHttps })
-
-  setCookie(event, jwtCookieName, jwt, {
-    expires: jwt ? dateInAMonth : dateEpoch,
-    httpOnly: true,
-    // path: '/',
-    sameSite: 'lax', // Cannot be 'strict' to allow authentications after clicking on links within webmailers.
-    secure: isHttps,
-  })
 }
