+++ src/shared/utils/api.ts
@@ -1,5 +1,4 @@
-import type { NuxtError } from '#app'
-import { consola } from 'consola'
+import type { AppGraphQLError, AppCombinedError } from '../types/api' // TODO: remove import once shared namespace has proper auto-imports (https://github.com/nuxt/nuxt/issues/32714)
 
 export const getGraphQLErrorMessage = ({
   graphqlError,
@@ -30,9 +29,4 @@ export const getCombinedErrorMessages = (
   }
 
   return errorMessages
-}
-
-export const showAppError = (error: string | Partial<NuxtError>) => {
-  showError(error)
-  consola.error(error)
 }
