--- src/shared/utils/auth.ts
@@ -1,5 +1,0 @@
-export const useJwtName = () => {
-  const { siteUrlTyped } = useSiteUrl()
-
-  return JWT_NAME({ isHttps: siteUrlTyped.protocol === 'https:' })
-}
