+++ src/shared/utils/dependencies/jose.ts
@@ -1,5 +1,7 @@
 import { jwtVerify, importSPKI } from 'jose'
 
+import { JWT_ALGORITHM } from '../constants' // TODO: remove import once shared namespace has proper auto-imports (https://github.com/nuxt/nuxt/issues/32714)
+
 export const verifyJwt = async ({
   jwt,
   jwtPublicKey,
