+++ src/tsconfig.json
@@ -5,13 +5,13 @@
       "path": "./.nuxt/tsconfig.app.json"
     },
     {
-      "path": "./.nuxt/tsconfig.node.json"
-    },
-    {
       "path": "./.nuxt/tsconfig.server.json"
     },
     {
       "path": "./.nuxt/tsconfig.shared.json"
+    },
+    {
+      "path": "./.nuxt/tsconfig.node.json"
     }
   ]
 }
