+++ tests/e2e/specs/server/headers.spec.ts
@@ -8,7 +8,7 @@ test.describe('headers middleware', () => {
 
     const HEADERS = {
       'access-control-allow-origin': '*',
-      'Content-Security-Policy': `base-uri 'none'; default-src 'none'; connect-src blob: https://localhost:3001 https://postgraphile.localhost:3001 https://tusd.localhost:3001 https://nominatim.openstreetmap.org/search https://cloudflareinsights.com https://*.analytics.google.com https://*.google-analytics.com https://*.googletagmanager.com https://o4507213726154752.ingest.de.sentry.io/api/4507213736837200/envelope/ 'self' https://firebaseinstallations.googleapis.com https://fcmregistrations.googleapis.com; font-src 'self' data:; form-action 'self'; frame-ancestors 'none'; frame-src https://challenges.cloudflare.com; img-src blob: https://tile.openstreetmap.org/ https://tusd.localhost:3001 https://www.gravatar.com/avatar/ https://*.google-analytics.com https://*.googletagmanager.com 'self' data:; manifest-src 'self'; style-src 'self' 'unsafe-inline'; script-src 'wasm-unsafe-eval'; worker-src http://localhost:3001/sw.js blob:; script-src-elem https://cdnjs.cloudflare.com/polyfill/v3/polyfill.min.js https://static.cloudflareinsights.com http://localhost:3001/cdn-cgi/ https://challenges.cloudflare.com https://*.googletagmanager.com 'nonce' http://localhost:3001/_nuxt/;`,
+      'Content-Security-Policy': `base-uri 'none'; default-src 'none'; connect-src blob: https://localhost:3001 https://postgraphile.localhost:3001 https://tusd.localhost:3001 https://nominatim.openstreetmap.org/search https://cloudflareinsights.com https://firebaseinstallations.googleapis.com https://fcmregistrations.googleapis.com https://*.analytics.google.com https://*.google-analytics.com https://*.googletagmanager.com https://o4507213726154752.ingest.de.sentry.io/api/4507213736837200/envelope/ 'self'; font-src 'self' data:; form-action 'self'; frame-ancestors 'none'; frame-src https://challenges.cloudflare.com; img-src blob: https://tile.openstreetmap.org/ https://tusd.localhost:3001 https://www.gravatar.com/avatar/ https://*.google-analytics.com https://*.googletagmanager.com 'self' data:; manifest-src 'self'; style-src 'self' 'unsafe-inline'; script-src 'wasm-unsafe-eval'; worker-src http://localhost:3001/sw.js blob:; script-src-elem https://cdnjs.cloudflare.com/polyfill/v3/polyfill.min.js https://static.cloudflareinsights.com http://localhost:3001/cdn-cgi/ https://challenges.cloudflare.com https://*.googletagmanager.com 'nonce' http://localhost:3001/_nuxt/;`,
       'content-type': 'text/html;charset=utf-8',
       'Cross-Origin-Embedder-Policy': 'credentialless',
       'Cross-Origin-Opener-Policy': 'same-origin',
